Move macro to its own module
This should make the diff a little nicer and wraps up all the doctests
into their own module nicely.
